Update ruleset.xml
Removes default arguments given to phpcs. See #5. This allows for more customization or manually passing preferred arguments to phpcs. This is kind of a breaking change because the default behavior changes. To run phpcs with the same arguments before version ^2.0 of this package, run:

php vendor/bin/phpcs -spvnl --standard=phpcs-laravel <files / directories>